# docs/API/openapi.yaml
openapi: 3.0.0
info:
  title: AI Assistant API
  version: 1.0.0
  description: |
    RAG-based policy assistant API for Caribbean government staff.

    ## Authentication
    All endpoints except `/branding` require authentication via NextAuth session cookies.

    ## User Roles
    - **admin**: Full system access
    - **superuser**: Manage documents/users in assigned categories
    - **user**: Query documents in subscribed categories
  contact:
    name: AI Assistant Support
  license:
    name: Proprietary
servers:
  - url: https://ai.abhirup.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          enum:
            - AUTH_REQUIRED
            - ADMIN_REQUIRED
            - SUPERUSER_REQUIRED
            - NOT_FOUND
            - FORBIDDEN
            - INVALID_BODY
            - INVALID_ID
            - INVALID_DAYS
            - VALIDATION
            - VALIDATION_ERROR
            - FILE_TOO_LARGE
            - UPLOAD_LIMIT
            - INVALID_FILE_TYPE
            - DUPLICATE
            - INTERNAL
            - SERVICE_ERROR
            - LEGACY_WRITE_DISABLED
        details:
          type: string
          description: Additional context
      required:
        - error
...
    AiSetupRedactedCredential:
      type: object
      description: Credential metadata with secret material redacted. Raw keys are never returned.
      required:
        - credentialId
        - providerId
        - status
        - isDefault
        - credentialVersion
        - lastVerifiedAt
        - redactedDetail
        - createdAt
        - updatedAt
      properties:
        credentialId:
          type: string
        providerId:
          type: string
        status:
          type: string
          example: active
        isDefault:
          type: boolean
        credentialVersion:
          type: integer
          minimum: 1
        lastVerifiedAt:
          type: string
          format: date-time
          nullable: true
        redactedDetail:
          type: string
          description: Bounded redaction placeholder; never contains the raw credential.
          example: '••••••••'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key for agent bot invocation
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Unauthorized"
            code: "AUTH_REQUIRED"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Admin access required"
            code: "ADMIN_REQUIRED"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Internal server error"
            code: "SERVICE_ERROR"
security:
  - sessionAuth: []
paths:
...
  /admin/ai-setup/organizations/{id}/credentials:
    parameters:
      - name: id
        in: path
        required: true
        description: Numeric organization ID. Server-side organization scope is always enforced.
        schema:
          type: integer
    get:
      tags:
        - Admin - AI & API Setup
      summary: List redacted organization credentials
      description: |
        Lists credential metadata for a visible organization. Secret ciphertext is not
        selected and raw credential values are never returned. Requires `super_admin`,
        or a global `admin` with a membership in its own organization.
      operationId: listAiSetupCredentials
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Redacted credential list
          content:
            application/json:
              schema:
                type: object
                required: [credentials]
                properties:
                  credentials:
                    type: array
                    items:
                      $ref: '#/components/schemas/AiSetupRedactedCredential'
        '400':
          description: Invalid organization ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Admin - AI & API Setup
      summary: Create or replace an organization credential
      description: |
        Creates a credential when `credentialId` is omitted, or replaces an existing
        credential for the same organization and provider when it is supplied. The
        submitted `secret` is encrypted by the credential vault and is never returned.
        Requires `super_admin` or the organization's server-resolved `org_admin`.
      operationId: upsertAiSetupCredential
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerId, secret]
              properties:
                providerId:
                  type: string
                secret:
                  type: string
                  format: password
                  writeOnly: true
                  minLength: 1
                  description: Raw credential accepted only on write; never echoed, logged, or returned.
                isDefault:
                  type: boolean
                  default: false
                credentialId:
                  type: string
                  description: Existing credential ID to replace. Omit to create a server-generated ID.
      responses:
        '200':
          description: Existing credential replaced
          content:
            application/json:
              schema:
                type: object
                required: [ok, action, credentialId, providerId]
                properties:
                  ok:
                    type: boolean
                    example: true
                  action:
                    type: string
                    enum: [replaced]
                  credentialId:
                    type: string
                  providerId:
                    type: string
        '201':
          description: Credential created
          content:
            application/json:
              schema:
                type: object
                required: [ok, action, credentialId, providerId]
                properties:
                  ok:
                    type: boolean
                    example: true
                  action:
                    type: string
                    enum: [created]
                  credentialId:
                    type: string
                  providerId:
                    type: string
        '400':
          description: Invalid ID/body, missing secret/provider, or unknown provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/ai-setup/organizations/{id}/members/{userId}:
    parameters:
      - name: id
        in: path
        required: true
        description: Numeric organization ID. Server-side organization scope is always enforced.
        schema:
          type: integer
      - name: userId
        in: path
        required: true
        schema:
          type: integer
    patch:
      tags:
        - Admin - AI & API Setup
      summary: Change an organization member role
      description: |
        Promoting a member delegates `org_admin` and demotes the previous active
        administrator. Demoting the last active `org_admin` is blocked. Requires
        `super_admin` or the organization's server-resolved `org_admin`.
      operationId: updateAiSetupMember
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [org_admin, member]
      responses:
        '200':
          description: Member role changed
          content:
            application/json:
              schema:
                type: object
                required: [ok, userId, role]
                properties:
                  ok:
                    type: boolean
                    example: true
                  userId:
                    type: integer
                  role:
                    type: string
                    enum: [org_admin, member]
        '400':
          description: Invalid ID/role or attempt to demote the last active administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Admin - AI & API Setup
      summary: Remove an organization member
      description: |
        Removes a member. Removing the last active `org_admin` is blocked until a
        successor is assigned. Requires `super_admin` or the organization's
        server-resolved `org_admin`.
      operationId: deleteAiSetupMember
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
                    example: true
        '400':
          description: Invalid ID or attempt to remove the last active administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...